finished unit tests
just need pargrp tests
