+++ src/testing/TestChronoTimer2.java
@@ -8,208 +8,122 @@ import org.junit.Test;
 
 import core.*;
 
-public class TestChronoTimer2 {
-	Chronotimer chronotimer = new Chronotimer();
-	
-	@Test
-	public void testReset() {
-		chronotimer.togglePower();
-		chronotimer.reset();
-		
-		assertEquals(true, chronotimer.getIsPoweredOn());
-		assertEquals(null, chronotimer.getCurrentRun());
-		assertEquals(RaceType.IND, chronotimer.getRaceType());
-	}
-	
-	@Test
-	public void testPower() {
-		assertEquals(false, chronotimer.getIsPoweredOn());
-		chronotimer.togglePower();
-		assertEquals(true, chronotimer.getIsPoweredOn());
-		chronotimer.togglePower();
-		assertEquals(false, chronotimer.getIsPoweredOn());
-	}
-	
-	@Test
-	public void testChannels() {
-		chronotimer.togglePower();
 
-		//default state is off
-		for(int i = 1; i < 9; i++){
-			assertTrue(!chronotimer.getChannelIsEnabled(i));
-			chronotimer.toggleChannel(i);
-		}
-		//turn all channels on
-		for(int i = 1; i < 9; i++){
-			assertTrue(chronotimer.getChannelIsEnabled(i));
-			chronotimer.toggleChannel(i);
-		}
-		//turn them all off again
-		for(int i = 1; i < 9; i++){
-			assertTrue(!chronotimer.getChannelIsEnabled(i));
-		}
-		//turn random channel on
-		chronotimer.toggleChannel(3);
-		assertTrue(chronotimer.getChannelIsEnabled(3));
-	}
+public class TestChronoTimer2 {
 	
 	@Test
-	public void testRun() {
-		chronotimer.togglePower();
+	public void testParInd() {
+		System.out.println("---Test ParInd---");
 		
+		/*
+		 * Unit tests that reflect some of the outputs
+		 * of chrono3.txt
+		 */
+		
+		Chronotimer chronotimer = new Chronotimer();
+		
+		//"POWER"
+		System.out.println("Turning on power...");
+		assertEquals(false, chronotimer.getIsPoweredOn());
+		chronotimer.togglePower();
+		assertEquals(true, chronotimer.getIsPoweredOn());
+		
+		//"TIME"
+		System.out.println("Setting time...");
+		chronotimer.setTime(12, 1, 30);
+		assertEquals("12:1:30.0", chronotimer.getTime());
+		
+		//"EVENT IND"
+		assertEquals(RaceType.IND, chronotimer.getRaceType());
+		chronotimer.setRaceType(RaceType.PARIND);
+		assertEquals(RaceType.PARIND, chronotimer.getRaceType());
+		System.out.println("Set race type: PARIND");
+				
+		//"NEWRUN"
+		assertEquals(null, chronotimer.getCurrentRun());
 		chronotimer.newRun();
 		assertNotEquals(null, chronotimer.getCurrentRun());
-		assertTrue(chronotimer.getCurrentRun() instanceof IndRun);
-		assertEquals(1, chronotimer.getRunNumber());
+		System.out.println("New run initiated");		
 		
+		System.out.println("First run test beginning...");
+		
+		//"TOG" - Toggle Channels 1 - 4
+		assertEquals(false, chronotimer.getChannelIsEnabled(1));
+		chronotimer.toggleChannel(1);
+		assertEquals(true, chronotimer.getChannelIsEnabled(1));
+		System.out.println("Toggled channel 1");
+		
+		assertEquals(false, chronotimer.getChannelIsEnabled(2));
+		chronotimer.toggleChannel(2);
+		assertEquals(true, chronotimer.getChannelIsEnabled(2));	
+		System.out.println("Toggled channel 2");
+		
+		assertEquals(false, chronotimer.getChannelIsEnabled(3));
+		chronotimer.toggleChannel(3);
+		assertEquals(true, chronotimer.getChannelIsEnabled(3));	
+		System.out.println("Toggled channel 3");
+		
+		assertEquals(false, chronotimer.getChannelIsEnabled(4));
+		chronotimer.toggleChannel(4);
+		assertEquals(true, chronotimer.getChannelIsEnabled(4));	
+		System.out.println("Toggled channel 4");
+
+		//"NUM" - Adds racer 234 and 315
+		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(272));
+		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(123));
+		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(111));
+		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(711));
+		chronotimer.addRacer(272);
+		chronotimer.addRacer(123);
+		chronotimer.addRacer(111);
+		chronotimer.addRacer(711);
+		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(272));
+		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(123));
+		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(111));
+		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(711));
+		System.out.println("Added racer (272)");
+		System.out.println("Added racer (123)");
+		System.out.println("Added racer (111)");
+		System.out.println("Added racer (711)");
+		
+		//"TRIG" Start
+		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(272));
+		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(111));
+		chronotimer.triggerChannel(1);
+		chronotimer.triggerChannel(3);
+		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(272));
+		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(123));
+		System.out.println("Racer (123) Triggered channel 1");
+		System.out.println("Racer (272) triggered channel 3");
+		
+		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(111));
+		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(711));
+		
+		//"TRIG" End
+		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInEndQueue(272));
+		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInEndQueue(123));
+		chronotimer.triggerChannel(2);
+		chronotimer.triggerChannel(4);
+		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInEndQueue(123));
+		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInEndQueue(272));
+		System.out.println("Racer (123) Triggered channel 2");
+		System.out.println("Racer (272) Triggered channel 4");
+		
+		assertTrue(chronotimer.getCurrentRun().getCurrentRunningRacers().length == 0);
+		
+		//"PRINT"
+		chronotimer.print();
+		
+		//"ENDRUN"
+		assertNotEquals(null, chronotimer.getCurrentRun());
 		chronotimer.endRun();
 		assertEquals(null, chronotimer.getCurrentRun());
-		
-		chronotimer.newRun();
-		assertEquals(2, chronotimer.getRunNumber());
-		
-	}
-	
-	@Test
-	public void testAddRacer() {
-		chronotimer.togglePower();
-		chronotimer.newRun();
-		chronotimer.addRacer(123);
-		
-		//prints error message
-		chronotimer.addRacer(123);
-		
-		chronotimer.addRacer(134);
-		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(123));
-		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(134));
-		
-		chronotimer.addRacer(175);
-		chronotimer.addRacer(189);
-		assertEquals(4, chronotimer.getCurrentRun().getCurrentWaitingRacers().length);
-	}
-	
-	@Test
-	public void testTrigger() {
-		chronotimer.togglePower();
-		chronotimer.newRun();
-		chronotimer.addRacer(123);
-		chronotimer.triggerChannel(1);
-		
-		assertEquals(0, chronotimer.getCurrentRun().getCurrentRunningRacers().length);
-		
-		chronotimer.toggleChannel(1);
-		chronotimer.triggerChannel(1);
-
-		assertEquals(1, chronotimer.getCurrentRun().getCurrentRunningRacers().length);
-		System.out.println("Racer 123 Start Time: " + chronotimer.getCurrentRun().getCurrentRunningRacers()[0].getStartTime());
-		
-		chronotimer.toggleChannel(2);
-		chronotimer.triggerChannel(2);
-		
-		assertEquals(1, chronotimer.getCurrentRun().getFinishedRacers().length);
-	}
-	
-	@Test
-	public void testSwap() {
-		chronotimer.togglePower();
-		chronotimer.setRaceType(RaceType.valueOf("IND"));
-		chronotimer.newRun();
+		System.out.println("Run ended");
 				
-		chronotimer.addRacer(10);
-		chronotimer.addRacer(11);
-		
-		chronotimer.toggleChannel(1);
-		chronotimer.toggleChannel(2);
-		assertTrue(chronotimer.getChannelIsEnabled(1));
-		assertTrue(chronotimer.getChannelIsEnabled(2));
-		
-		chronotimer.triggerChannel(1);
-		chronotimer.triggerChannel(1);
-		
-		chronotimer.swap();
-		
-		assertEquals(11, chronotimer.getCurrentRun().getCurrentRunningRacers()[0].getBibNumber());
-	}
-	
-	@Test
-	public void testSwap2(){
+		//"POWER"
+		assertEquals(true, chronotimer.getIsPoweredOn());
 		chronotimer.togglePower();
-		chronotimer.setRaceType(RaceType.valueOf("IND"));
-		chronotimer.newRun();
-				
-		chronotimer.addRacer(10);
-		chronotimer.addRacer(11);
-		chronotimer.addRacer(12);
-		chronotimer.addRacer(13);
-		
-		chronotimer.toggleChannel(1);
-		chronotimer.toggleChannel(2);
-		assertTrue(chronotimer.getChannelIsEnabled(1));
-		assertTrue(chronotimer.getChannelIsEnabled(2));
-		
-		chronotimer.triggerChannel(1);
-		chronotimer.triggerChannel(1);
-		chronotimer.triggerChannel(1);
-		chronotimer.triggerChannel(1);
-		
-		chronotimer.swap();
-		
-		assertEquals(11, chronotimer.getCurrentRun().getCurrentRunningRacers()[0].getBibNumber());
-		
-		chronotimer.triggerChannel(2);
-		chronotimer.triggerChannel(2);
-		
-		chronotimer.swap();
-		
-		assertEquals(13, chronotimer.getCurrentRun().getCurrentRunningRacers()[0].getBibNumber());
-		
-		chronotimer.triggerChannel(2);
-		chronotimer.triggerChannel(2);
-		
-		assertEquals(0, chronotimer.getCurrentRun().getCurrentRunningRacers().length);
-	}
-	
-	@Test
-	public void testCancelIND() {
-		chronotimer.togglePower();
-		chronotimer.setRaceType(RaceType.IND);
-		chronotimer.newRun();
-		chronotimer.addRacer(10);
-		chronotimer.toggleChannel(1);
-		
-		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(10));
-		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(10));
-		
-		chronotimer.triggerChannel(1);
-		
-		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(10));
-		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(10));
-		
-		chronotimer.cancel();
-		
-		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(10));
-		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(10));
-	}
-	
-	@Test
-	public void testCancelPARIND() {
-		chronotimer.togglePower();
-		chronotimer.setRaceType(RaceType.PARIND);
-		chronotimer.newRun();
-		chronotimer.addRacer(10);
-		chronotimer.toggleChannel(1);
-		
-		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(10));
-		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(10));
-		
-		chronotimer.triggerChannel(1);
-		
-		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(10));
-		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(10));
-		
-		chronotimer.cancel();
-		
-		assertEquals(true, chronotimer.getCurrentRun().containsRacerBibNumberInWaitQueue(10));
-		assertEquals(false, chronotimer.getCurrentRun().containsRacerBibNumberInRunningQueue(10));
+		assertEquals(false, chronotimer.getIsPoweredOn());
+		System.out.println("Turning off power...");
 	}
 }
